Removed duplication where possible in the analytics and scoped SQL components and backend.
