+++ Tombolo/client-reactjs/src/services/workunitAnalytics.service.ts
@@ -67,38 +67,6 @@ const analyticsService = {
     return response.data;
   },
 
-  exportToCSV: (results: { columns: string[]; rows: any[] }, filename = 'query-results'): void => {
-    if (!results || !results.rows || results.rows.length === 0) {
-      throw new Error('No data to export');
-    }
-
-    const headers = results.columns.join(',');
-    const rows = results.rows
-      .map(row =>
-        results.columns
-          .map(col => {
-            const value = row[col];
-            if (value === null || value === undefined) return '';
-            const stringValue = String(value);
-            if (stringValue.includes(',') || stringValue.includes('"') || stringValue.includes('\n')) {
-              return `"${stringValue.replace(/"/g, '""')}"`;
-            }
-            return stringValue;
-          })
-          .join(',')
-      )
-      .join('\n');
-
-    const csv = `${headers}\n${rows}`;
-    const blob = new Blob([csv], { type: 'text/csv;charset=utf-8;' });
-    const url = URL.createObjectURL(blob);
-    const link = document.createElement('a');
-    link.href = url;
-    link.download = `${filename}-${Date.now()}.csv`;
-    link.click();
-    URL.revokeObjectURL(url);
-  },
-
   exportToJSON: (results: { rows: any[] }, filename = 'query-results'): void => {
     if (!results || !results.rows || results.rows.length === 0) {
       throw new Error('No data to export');
